server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
dj-database-url==2.1.0
django-filter==24.3